Move vision tests out to own file
